+++ src/DataFixtures/Files/Suivi.yml
@@ -91,3 +91,27 @@ suivis:
   is_public: 0
   signalement: "2023-6"
   type: 1
+-
+  created_by: admin-01@histologe.fr
+  description: "Signalement validé"
+  is_public: 1
+  signalement: "2023-7"
+  type: 1
+-
+  created_by: admin-01@histologe.fr
+  description: "Signalement validé"
+  is_public: 1
+  signalement: "2023-8"
+  type: 1
+-
+  created_by: admin-01@histologe.fr
+  description: "Signalement validé"
+  is_public: 1
+  signalement: "2023-9"
+  type: 1
+-
+  created_by: admin-01@histologe.fr
+  description: "Signalement validé"
+  is_public: 1
+  signalement: "2023-10"
+  type: 1
